# Conflicts:
#	CHANGELOG.md
#	src/main/bundle/version

+++ src/main/bundle/codedeploy-common-files/decrypt-properties.bash
@@ -6,5 +6,9 @@ cd ${SCRIPT_PATH}
 
 for file in $(find ./data -name '*.properties')
 do
+  if [ ! -f ${file}.encrypted ]; then
     cat ${file} | python decrypt-file.py > ${file}.decrypted
+    mv ${file} ${file}.encrypted
+    mv ${file}.decrypted ${file}
+  fi
 done
